% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My binomial simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of trials to run}

\item{n}{Size of sample}

\item{p}{Probability of success (1)}
}
\value{
A table and barplot with the simulated frequency distribution
}
\description{
My binomial simulation
}
\examples{
mybin(iter=10000, n=20, p=0.5)

}
