% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My normal curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The area of the distribution under X≤a will be shaded}
}
\value{
A plot of the normal distribution with with the area under X≤a shaded; the area
}
\description{
My normal curve
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
