% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Outliers}
\usage{
outliers(data)
}
\arguments{
\item{data}{A quantitative vector}
}
\value{
A list containing possible and defined outliers. Possible outliers have a z-score between 2 and 3, while defined outliers have a z-score greater than 3.
}
\description{
Outliers
}
\examples{
d <- 1:40; outliers(data=d)
}
